scripts/mlflow_migration/requirements.txt
clearml>=0.16.3,<0.18.0
mlflow==1.12.1
pandas
pillow
python-dateutil
pyyaml
tqdm
